--- Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Config/SecurityConfig.java
@@ -1,47 +1,0 @@
-package com.example.springbootmonolith.Config;
-
-import com.example.springbootmonolith.Service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    UserService userService;
-
-    @Autowired
-    private JWTRequestFilter jwtRequestFilter;
-
-    @Bean("encoder")
-    public PasswordEncoder encoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-//    public void configureGlobalSecurity(AuthenticationManagerBuilder auth) throws Exception {
-//        auth.userDetailsService(userService);
-//    }
-@Override
-    protected void configure(HttpSecurity http) throws Exception {
-    http.csrf().disable()
-            .authorizeRequests()
-            .antMatchers("/signup/**", "/login/**").permitAll()
-            .antMatchers("/user/**", "/profile/**", "/comment/**", "/post/**").authenticated()
-            .antMatchers("/role/**").hasRole("DBA")
-            .and()
-            .httpBasic();
-
-    http.addFilterBefore(jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
-
-}
-
-
-}
