+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.example.springbootmonolith.Repository;
 
-import com.example.springbootmonolith.Model.Post;
+import com.example.springbootmonolith.model.Post;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
