+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/CommentService.java
@@ -1,6 +1,6 @@
 package com.example.springbootmonolith.Service;
 
-import com.example.springbootmonolith.Model.Comment;
+import com.example.springbootmonolith.model.Comment;
 
 public interface CommentService {
