+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/CommentServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.springbootmonolith.Service;
 
 import com.example.springbootmonolith.Repository.CommentRepository;
-import com.example.springbootmonolith.Model.Comment;
+import com.example.springbootmonolith.model.Comment;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class CommentServiceImpl implements CommentService {
