+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/PostServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.springbootmonolith.Service;
 
 import com.example.springbootmonolith.Repository.PostRepository;
-import com.example.springbootmonolith.Model.Post;
+import com.example.springbootmonolith.model.Post;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
