+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/UserProfileService.java
@@ -1,6 +1,6 @@
 package com.example.springbootmonolith.Service;
 
-import com.example.springbootmonolith.Model.UserProfile;
+import com.example.springbootmonolith.model.UserProfile;
 
 public interface UserProfileService {
