+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/UserProfileServiceImpl.java
@@ -1,8 +1,8 @@
 package com.example.springbootmonolith.Service;
 
 import com.example.springbootmonolith.Repository.UserProfileRepository;
-import com.example.springbootmonolith.Model.User;
-import com.example.springbootmonolith.Model.UserProfile;
+import com.example.springbootmonolith.model.User;
+import com.example.springbootmonolith.model.UserProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
