+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/UserServiceImpl.java
@@ -3,8 +3,10 @@ package com.example.springbootmonolith.Service;
 import com.example.springbootmonolith.Repository.CommentRepository;
 import com.example.springbootmonolith.Repository.PostRepository;
 import com.example.springbootmonolith.Repository.UserRepository;
-import com.example.springbootmonolith.Model.User;
+import com.example.springbootmonolith.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
